update : some minor issues

+++ src/PolarisPolicy.cc
@@ -59,7 +59,7 @@ PolarisPolicyConfig::PolarisPolicyConfig(const std::string& policy_name,
 void PolarisPolicyConfig::set_nearby_based_router(bool enable,
 												 const std::string& match_level,
 												 const std::string& max_match_level,
-												 short percentage,
+												 int percentage,
 												 bool enable_recover_all,
 												 bool strict_nearby)
 {
